@@ -1,8 +1,12 @@
 package resolvers
 
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+// Code generated by github.com/99designs/gqlgen version v0.17.57
+
 import (
 	"context"
-	"strings"
+	"fmt"
 
 	"github.com/photoview/photoview/api/dataloader"
 	api "github.com/photoview/photoview/api/graphql"
@@ -10,91 +14,34 @@ import (
 	"github.com/photoview/photoview/api/graphql/models"
 	"github.com/photoview/photoview/api/graphql/models/actions"
 	"github.com/photoview/photoview/api/scanner/face_detection"
-	"github.com/pkg/errors"
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
 )
 
-func (r *queryResolver) MyMedia(ctx context.Context, order *models.Ordering, paginate *models.Pagination) ([]*models.Media, error) {
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, errors.New("unauthorized")
-	}
-
-	return actions.MyMedia(r.DB(ctx), user, order, paginate)
+// Thumbnail is the resolver for the thumbnail field.
+func (r *mediaResolver) Thumbnail(ctx context.Context, obj *models.Media) (*models.MediaURL, error) {
+	return dataloader.For(ctx).MediaThumbnail.Load(obj.ID)
 }
 
-func (r *queryResolver) Media(ctx context.Context, id int, tokenCredentials *models.ShareTokenCredentials) (*models.Media, error) {
-	db := r.DB(ctx)
-	if tokenCredentials != nil {
-
-		shareToken, err := r.ShareToken(ctx, *tokenCredentials)
-		if err != nil {
-			return nil, err
-		}
-
-		if *shareToken.MediaID == id {
-			return shareToken.Media, nil
-		}
+// HighRes is the resolver for the highRes field.
+func (r *mediaResolver) HighRes(ctx context.Context, obj *models.Media) (*models.MediaURL, error) {
+	if obj.Type != models.MediaTypePhoto {
+		return nil, nil
 	}
 
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, auth.ErrUnauthorized
-	}
-
-	var media models.Media
-
-	err := db.
-		Joins("Album").
-		Where("media.id = ?", id).
-		Where("EXISTS (SELECT * FROM user_albums WHERE user_albums.album_id = media.album_id AND user_albums.user_id = ?)", user.ID).
-		Where("media.id IN (?)", db.Model(&models.MediaURL{}).Select("media_id").Where("media_urls.media_id = media.id")).
-		First(&media).Error
-
-	if err != nil {
-		return nil, errors.Wrap(err, "could not get media by media_id and user_id from database")
-	}
-
-	return &media, nil
+	return dataloader.For(ctx).MediaHighres.Load(obj.ID)
 }
 
-func (r *queryResolver) MediaList(ctx context.Context, ids []int) ([]*models.Media, error) {
-	db := r.DB(ctx)
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, auth.ErrUnauthorized
+// VideoWeb is the resolver for the videoWeb field.
+func (r *mediaResolver) VideoWeb(ctx context.Context, obj *models.Media) (*models.MediaURL, error) {
+	if obj.Type != models.MediaTypeVideo {
+		return nil, nil
 	}
 
-	if len(ids) == 0 {
-		return nil, errors.New("no ids provided")
-	}
-
-	var media []*models.Media
-	err := db.Model(&media).
-		Joins("LEFT JOIN user_albums ON user_albums.album_id = media.album_id").
-		Where("media.id IN ?", ids).
-		Where("user_albums.user_id = ?", user.ID).
-		Find(&media).Error
-
-	if err != nil {
-		return nil, errors.Wrap(err, "could not get media list by media_id and user_id from database")
-	}
-
-	return media, nil
+	return dataloader.For(ctx).MediaVideoWeb.Load(obj.ID)
 }
 
-type mediaResolver struct {
-	*Resolver
-}
-
-func (r *Resolver) Media() api.MediaResolver {
-	return &mediaResolver{r}
-}
-
-func (r *mediaResolver) Type(ctx context.Context, media *models.Media) (models.MediaType, error) {
-	formattedType := models.MediaType(strings.Title(string(media.Type)))
-	return formattedType, nil
-}
-
+// Album is the resolver for the album field.
 func (r *mediaResolver) Album(ctx context.Context, obj *models.Media) (*models.Album, error) {
 	var album models.Album
 	err := r.DB(ctx).Find(&album, obj.AlbumID).Error
@@ -104,20 +51,54 @@ func (r *mediaResolver) Album(ctx context.Context, obj *models.Media) (*models.A
 	return &album, nil
 }
 
-func (r *mediaResolver) Shares(ctx context.Context, media *models.Media) ([]*models.ShareToken, error) {
+// Exif is the resolver for the exif field.
+func (r *mediaResolver) Exif(ctx context.Context, obj *models.Media) (*models.MediaEXIF, error) {
+	if obj.Exif != nil {
+		return obj.Exif, nil
+	}
+
+	var exif models.MediaEXIF
+	if err := r.DB(ctx).Model(obj).Association("Exif").Find(&exif); err != nil {
+		return nil, err
+	}
+
+	return &exif, nil
+}
+
+// Favorite is the resolver for the favorite field.
+func (r *mediaResolver) Favorite(ctx context.Context, obj *models.Media) (bool, error) {
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return false, auth.ErrUnauthorized
+	}
+
+	return dataloader.For(ctx).UserMediaFavorite.Load(&models.UserMediaData{
+		UserID:  user.ID,
+		MediaID: obj.ID,
+	})
+}
+
+// Type is the resolver for the type field.
+func (r *mediaResolver) Type(ctx context.Context, obj *models.Media) (models.MediaType, error) {
+	formattedType := models.MediaType(cases.Title(language.Und).String(string(obj.Type)))
+	return formattedType, nil
+}
+
+// Shares is the resolver for the shares field.
+func (r *mediaResolver) Shares(ctx context.Context, obj *models.Media) ([]*models.ShareToken, error) {
 	var shareTokens []*models.ShareToken
-	if err := r.DB(ctx).Where("media_id = ?", media.ID).Find(&shareTokens).Error; err != nil {
-		return nil, errors.Wrapf(err, "get shares for media (%s)", media.Path)
+	if err := r.DB(ctx).Where("media_id = ?", obj.ID).Find(&shareTokens).Error; err != nil {
+		return nil, fmt.Errorf("get shares for media (%s): %w", obj.Path, err)
 	}
 
 	return shareTokens, nil
 }
 
-func (r *mediaResolver) Downloads(ctx context.Context, media *models.Media) ([]*models.MediaDownload, error) {
-
+// Downloads is the resolver for the downloads field.
+func (r *mediaResolver) Downloads(ctx context.Context, obj *models.Media) ([]*models.MediaDownload, error) {
 	var mediaUrls []*models.MediaURL
-	if err := r.DB(ctx).Where("media_id = ?", media.ID).Find(&mediaUrls).Error; err != nil {
-		return nil, errors.Wrapf(err, "get downloads for media (%s)", media.Path)
+	if err := r.DB(ctx).Where("media_id = ?", obj.ID).Find(&mediaUrls).Error; err != nil {
+		return nil, fmt.Errorf("get downloads for media (%s): %w", obj.Path, err)
 	}
 
 	downloads := make([]*models.MediaDownload, 0)
@@ -147,53 +128,26 @@ func (r *mediaResolver) Downloads(ctx context.Context, media *models.Media) ([]*
 	return downloads, nil
 }
 
-func (r *mediaResolver) HighRes(ctx context.Context, media *models.Media) (*models.MediaURL, error) {
-	if media.Type != models.MediaTypePhoto {
-		return nil, nil
+// Faces is the resolver for the faces field.
+func (r *mediaResolver) Faces(ctx context.Context, obj *models.Media) ([]*models.ImageFace, error) {
+	if face_detection.GlobalFaceDetector == nil {
+		return []*models.ImageFace{}, nil
 	}
 
-	return dataloader.For(ctx).MediaHighres.Load(media.ID)
-}
-
-func (r *mediaResolver) Thumbnail(ctx context.Context, media *models.Media) (*models.MediaURL, error) {
-	return dataloader.For(ctx).MediaThumbnail.Load(media.ID)
-}
-
-func (r *mediaResolver) VideoWeb(ctx context.Context, media *models.Media) (*models.MediaURL, error) {
-	if media.Type != models.MediaTypeVideo {
-		return nil, nil
+	if obj.Faces != nil {
+		return obj.Faces, nil
 	}
 
-	return dataloader.For(ctx).MediaVideoWeb.Load(media.ID)
-}
-
-func (r *mediaResolver) Exif(ctx context.Context, media *models.Media) (*models.MediaEXIF, error) {
-	if media.Exif != nil {
-		return media.Exif, nil
-	}
-
-	var exif models.MediaEXIF
-	if err := r.DB(ctx).Model(&media).Association("Exif").Find(&exif); err != nil {
+	var faces []*models.ImageFace
+	if err := r.DB(ctx).Model(obj).Association("Faces").Find(&faces); err != nil {
 		return nil, err
 	}
 
-	return &exif, nil
+	return faces, nil
 }
 
-func (r *mediaResolver) Favorite(ctx context.Context, media *models.Media) (bool, error) {
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return false, auth.ErrUnauthorized
-	}
-
-	return dataloader.For(ctx).UserMediaFavorite.Load(&models.UserMediaData{
-		UserID:  user.ID,
-		MediaID: media.ID,
-	})
-}
-
+// FavoriteMedia is the resolver for the favoriteMedia field.
 func (r *mutationResolver) FavoriteMedia(ctx context.Context, mediaID int, favorite bool) (*models.Media, error) {
-
 	user := auth.UserFromContext(ctx)
 	if user == nil {
 		return nil, auth.ErrUnauthorized
@@ -202,19 +156,80 @@ func (r *mutationResolver) FavoriteMedia(ctx context.Context, mediaID int, favor
 	return user.FavoriteMedia(r.DB(ctx), mediaID, favorite)
 }
 
-func (r *mediaResolver) Faces(ctx context.Context, media *models.Media) ([]*models.ImageFace, error) {
-	if face_detection.GlobalFaceDetector == nil {
-		return []*models.ImageFace{}, nil
+// MyMedia is the resolver for the myMedia field.
+func (r *queryResolver) MyMedia(ctx context.Context, order *models.Ordering, paginate *models.Pagination) ([]*models.Media, error) {
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, fmt.Errorf("unauthorized")
 	}
 
-	if media.Faces != nil {
-		return media.Faces, nil
-	}
-
-	var faces []*models.ImageFace
-	if err := r.DB(ctx).Model(&media).Association("Faces").Find(&faces); err != nil {
-		return nil, err
-	}
-
-	return faces, nil
+	return actions.MyMedia(r.DB(ctx), user, order, paginate)
 }
+
+// Media is the resolver for the media field.
+func (r *queryResolver) Media(ctx context.Context, id int, tokenCredentials *models.ShareTokenCredentials) (*models.Media, error) {
+	db := r.DB(ctx)
+	if tokenCredentials != nil {
+
+		shareToken, err := r.ShareToken(ctx, *tokenCredentials)
+		if err != nil {
+			return nil, err
+		}
+
+		if *shareToken.MediaID == id {
+			return shareToken.Media, nil
+		}
+	}
+
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, auth.ErrUnauthorized
+	}
+
+	var media models.Media
+
+	err := db.
+		Joins("Album").
+		Where("media.id = ?", id).
+		Where("EXISTS (SELECT * FROM user_albums WHERE user_albums.album_id = media.album_id AND user_albums.user_id = ?)",
+			user.ID).
+		Where("media.id IN (?)", db.Model(&models.MediaURL{}).Select("media_id").Where("media_urls.media_id = media.id")).
+		First(&media).Error
+
+	if err != nil {
+		return nil, fmt.Errorf("could not get media by media_id and user_id from database: %w", err)
+	}
+
+	return &media, nil
+}
+
+// MediaList is the resolver for the mediaList field.
+func (r *queryResolver) MediaList(ctx context.Context, ids []int) ([]*models.Media, error) {
+	db := r.DB(ctx)
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, auth.ErrUnauthorized
+	}
+
+	if len(ids) == 0 {
+		return nil, fmt.Errorf("no ids provided")
+	}
+
+	var media []*models.Media
+	err := db.Model(&media).
+		Joins("LEFT JOIN user_albums ON user_albums.album_id = media.album_id").
+		Where("media.id IN ?", ids).
+		Where("user_albums.user_id = ?", user.ID).
+		Find(&media).Error
+
+	if err != nil {
+		return nil, fmt.Errorf("could not get media list by media_id and user_id from database: %w", err)
+	}
+
+	return media, nil
+}
+
+// Media returns api.MediaResolver implementation.
+func (r *Resolver) Media() api.MediaResolver { return &mediaResolver{r} }
+
+type mediaResolver struct{ *Resolver }
