@@ -1,77 +1,25 @@
 package resolvers
 
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+// Code generated by github.com/99designs/gqlgen version v0.17.57
+
 import (
 	"context"
-	"os"
+	"errors"
+	"fmt"
 	"path"
-	"strconv"
 
 	api "github.com/photoview/photoview/api/graphql"
 	"github.com/photoview/photoview/api/graphql/auth"
 	"github.com/photoview/photoview/api/graphql/models"
 	"github.com/photoview/photoview/api/graphql/models/actions"
 	"github.com/photoview/photoview/api/scanner"
-	"github.com/photoview/photoview/api/scanner/face_detection"
-	"github.com/photoview/photoview/api/utils"
-	"github.com/pkg/errors"
 	"golang.org/x/crypto/bcrypt"
 	"gorm.io/gorm"
 )
 
-type userResolver struct {
-	*Resolver
-}
-
-func (r *Resolver) User() api.UserResolver {
-	return &userResolver{r}
-}
-
-func (r *queryResolver) User(ctx context.Context, order *models.Ordering, paginate *models.Pagination) ([]*models.User, error) {
-
-	var users []*models.User
-
-	if err := models.FormatSQL(r.DB(ctx).Model(models.User{}), order, paginate).Find(&users).Error; err != nil {
-		return nil, err
-	}
-
-	return users, nil
-}
-
-func (r *userResolver) Albums(ctx context.Context, user *models.User) ([]*models.Album, error) {
-	user.FillAlbums(r.DB(ctx))
-
-	pointerAlbums := make([]*models.Album, len(user.Albums))
-	for i, album := range user.Albums {
-		pointerAlbums[i] = &album
-	}
-
-	return pointerAlbums, nil
-}
-
-func (r *userResolver) RootAlbums(ctx context.Context, user *models.User) (albums []*models.Album, err error) {
-	db := r.DB(ctx)
-
-	err = db.Model(&user).
-		Where("albums.parent_album_id NOT IN (?)",
-			db.Table("user_albums").
-				Select("albums.id").
-				Joins("JOIN albums ON albums.id = user_albums.album_id AND user_albums.user_id = ?", user.ID),
-		).Or("albums.parent_album_id IS NULL").
-		Association("Albums").Find(&albums)
-
-	return
-}
-
-func (r *queryResolver) MyUser(ctx context.Context) (*models.User, error) {
-
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, auth.ErrUnauthorized
-	}
-
-	return user, nil
-}
-
+// AuthorizeUser is the resolver for the authorizeUser field.
 func (r *mutationResolver) AuthorizeUser(ctx context.Context, username string, password string) (*models.AuthorizeResult, error) {
 	db := r.DB(ctx)
 	user, err := models.AuthorizeUser(db, username, password)
@@ -104,6 +52,7 @@ func (r *mutationResolver) AuthorizeUser(ctx context.Context, username string, p
 	}, nil
 }
 
+// InitialSetupWizard is the resolver for the initialSetupWizard field.
 func (r *mutationResolver) InitialSetupWizard(ctx context.Context, username string, password string, rootPath string) (*models.AuthorizeResult, error) {
 	db := r.DB(ctx)
 	siteInfo, err := models.GetSiteInfo(db)
@@ -156,51 +105,7 @@ func (r *mutationResolver) InitialSetupWizard(ctx context.Context, username stri
 	}, nil
 }
 
-func (r *queryResolver) MyUserPreferences(ctx context.Context) (*models.UserPreferences, error) {
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, auth.ErrUnauthorized
-	}
-
-	userPref := models.UserPreferences{
-		UserID: user.ID,
-	}
-	if err := r.DB(ctx).Where("user_id = ?", user.ID).FirstOrCreate(&userPref).Error; err != nil {
-		return nil, err
-	}
-
-	return &userPref, nil
-}
-
-func (r *mutationResolver) ChangeUserPreferences(ctx context.Context, language *string) (*models.UserPreferences, error) {
-	db := r.DB(ctx)
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, auth.ErrUnauthorized
-	}
-
-	var langTrans *models.LanguageTranslation = nil
-	if language != nil {
-		lng := models.LanguageTranslation(*language)
-		langTrans = &lng
-	}
-
-	var userPref models.UserPreferences
-	if err := db.Where("user_id = ?", user.ID).FirstOrInit(&userPref).Error; err != nil {
-		return nil, err
-	}
-
-	userPref.UserID = user.ID
-	userPref.Language = langTrans
-
-	if err := db.Save(&userPref).Error; err != nil {
-		return nil, err
-	}
-
-	return &userPref, nil
-}
-
-// Admin queries
+// UpdateUser is the resolver for the updateUser field.
 func (r *mutationResolver) UpdateUser(ctx context.Context, id int, username *string, password *string, admin *bool) (*models.User, error) {
 	db := r.DB(ctx)
 
@@ -232,14 +137,14 @@ func (r *mutationResolver) UpdateUser(ctx context.Context, id int, username *str
 	}
 
 	if err := db.Save(&user).Error; err != nil {
-		return nil, errors.Wrap(err, "failed to update user")
+		return nil, fmt.Errorf("failed to update user: %w", err)
 	}
 
 	return &user, nil
 }
 
+// CreateUser is the resolver for the createUser field.
 func (r *mutationResolver) CreateUser(ctx context.Context, username string, password *string, admin bool) (*models.User, error) {
-
 	var user *models.User
 
 	transactionError := r.DB(ctx).Transaction(func(tx *gorm.DB) error {
@@ -259,10 +164,12 @@ func (r *mutationResolver) CreateUser(ctx context.Context, username string, pass
 	return user, nil
 }
 
+// DeleteUser is the resolver for the deleteUser field.
 func (r *mutationResolver) DeleteUser(ctx context.Context, id int) (*models.User, error) {
 	return actions.DeleteUser(r.DB(ctx), id)
 }
 
+// UserAddRootPath is the resolver for the userAddRootPath field.
 func (r *mutationResolver) UserAddRootPath(ctx context.Context, id int, rootPath string) (*models.Album, error) {
 	db := r.DB(ctx)
 
@@ -281,6 +188,7 @@ func (r *mutationResolver) UserAddRootPath(ctx context.Context, id int, rootPath
 	return newAlbum, nil
 }
 
+// UserRemoveRootAlbum is the resolver for the userRemoveRootAlbum field.
 func (r *mutationResolver) UserRemoveRootAlbum(ctx context.Context, userID int, albumID int) (*models.Album, error) {
 	db := r.DB(ctx)
 
@@ -290,7 +198,6 @@ func (r *mutationResolver) UserRemoveRootAlbum(ctx context.Context, userID int, 
 	}
 
 	var deletedAlbumIDs []int = nil
-
 	transactionError := db.Transaction(func(tx *gorm.DB) error {
 		if err := tx.Raw("DELETE FROM user_albums WHERE user_id = ? AND album_id = ?", userID, albumID).Error; err != nil {
 			return err
@@ -316,20 +223,9 @@ func (r *mutationResolver) UserRemoveRootAlbum(ctx context.Context, userID int, 
 		}
 
 		// Cleanup if no user owns the album anymore
-		var userAlbumCount int
-		if err := tx.Raw("SELECT COUNT(user_id) FROM user_albums WHERE album_id = ?", albumID).Scan(&userAlbumCount).Error; err != nil {
+		deletedAlbumIDs, err = cleanup(tx, albumID, childAlbumIDs)
+		if err != nil {
 			return err
-		}
-
-		if userAlbumCount == 0 {
-			deletedAlbumIDs = append(childAlbumIDs, albumID)
-			childAlbumIDs = nil
-
-			// Delete albums from database
-			if err := tx.Delete(&models.Album{}, "id IN (?)", deletedAlbumIDs).Error; err != nil {
-				deletedAlbumIDs = nil
-				return err
-			}
 		}
 
 		return nil
@@ -339,23 +235,108 @@ func (r *mutationResolver) UserRemoveRootAlbum(ctx context.Context, userID int, 
 		return nil, transactionError
 	}
 
-	if deletedAlbumIDs != nil {
-		// Delete albums from cache
-		for _, id := range deletedAlbumIDs {
-			cacheAlbumPath := path.Join(utils.MediaCachePath(), strconv.Itoa(id))
-
-			if err := os.RemoveAll(cacheAlbumPath); err != nil {
-				return nil, err
-			}
-		}
-
-		// Reload faces as media might have been deleted
-		if face_detection.GlobalFaceDetector != nil {
-			if err := face_detection.GlobalFaceDetector.ReloadFacesFromDatabase(db); err != nil {
-				return nil, err
-			}
-		}
+	if err := clearCacheAndReloadFaces(db, deletedAlbumIDs); err != nil {
+		return nil, err
 	}
 
 	return &album, nil
 }
+
+// ChangeUserPreferences is the resolver for the changeUserPreferences field.
+func (r *mutationResolver) ChangeUserPreferences(ctx context.Context, language *string) (*models.UserPreferences, error) {
+	db := r.DB(ctx)
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, auth.ErrUnauthorized
+	}
+
+	var langTrans *models.LanguageTranslation = nil
+	if language != nil {
+		lng := models.LanguageTranslation(*language)
+		langTrans = &lng
+	}
+
+	var userPref models.UserPreferences
+	if err := db.Where("user_id = ?", user.ID).FirstOrInit(&userPref).Error; err != nil {
+		return nil, err
+	}
+
+	userPref.UserID = user.ID
+	userPref.Language = langTrans
+
+	if err := db.Save(&userPref).Error; err != nil {
+		return nil, err
+	}
+
+	return &userPref, nil
+}
+
+// User is the resolver for the user field.
+func (r *queryResolver) User(ctx context.Context, order *models.Ordering, paginate *models.Pagination) ([]*models.User, error) {
+	var users []*models.User
+
+	if err := models.FormatSQL(r.DB(ctx).Model(models.User{}), order, paginate).Find(&users).Error; err != nil {
+		return nil, err
+	}
+
+	return users, nil
+}
+
+// MyUser is the resolver for the myUser field.
+func (r *queryResolver) MyUser(ctx context.Context) (*models.User, error) {
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, auth.ErrUnauthorized
+	}
+
+	return user, nil
+}
+
+// MyUserPreferences is the resolver for the myUserPreferences field.
+func (r *queryResolver) MyUserPreferences(ctx context.Context) (*models.UserPreferences, error) {
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, auth.ErrUnauthorized
+	}
+
+	userPref := models.UserPreferences{
+		UserID: user.ID,
+	}
+	if err := r.DB(ctx).Where("user_id = ?", user.ID).FirstOrCreate(&userPref).Error; err != nil {
+		return nil, err
+	}
+
+	return &userPref, nil
+}
+
+// Albums is the resolver for the albums field.
+func (r *userResolver) Albums(ctx context.Context, obj *models.User) ([]*models.Album, error) {
+	obj.FillAlbums(r.DB(ctx))
+
+	pointerAlbums := make([]*models.Album, len(obj.Albums))
+	for i, album := range obj.Albums {
+		pointerAlbums[i] = &album
+	}
+
+	return pointerAlbums, nil
+}
+
+// RootAlbums is the resolver for the rootAlbums field.
+func (r *userResolver) RootAlbums(ctx context.Context, obj *models.User) (albums []*models.Album, err error) {
+	db := r.DB(ctx)
+
+	err = db.Model(obj).
+		Where("albums.parent_album_id NOT IN (?)",
+			db.Table("user_albums").
+				Select("albums.id").
+				Joins("JOIN albums ON albums.id = user_albums.album_id AND user_albums.user_id = ?", obj.ID),
+		).Or("albums.parent_album_id IS NULL").Order("path ASC").
+		Association("Albums").Find(&albums)
+
+	return
+}
+
+// User returns api.UserResolver implementation.
+func (r *Resolver) User() api.UserResolver { return &userResolver{r} }
+
+type userResolver struct{ *Resolver }
