@@ -8,7 +8,10 @@ import (
 	"gorm.io/gorm"
 )
 
-func MyTimeline(db *gorm.DB, user *models.User, paginate *models.Pagination, onlyFavorites *bool, fromDate *time.Time) ([]*models.Media, error) {
+func MyTimeline(db *gorm.DB, user *models.User, paginate *models.Pagination, onlyFavorites *bool,
+	fromDate *time.Time) ([]*models.Media, error) {
+
+	const albumsTitleASC = "albums.title ASC"
 
 	query := db.
 		Joins("JOIN albums ON media.album_id = albums.id").
@@ -20,19 +23,19 @@ func MyTimeline(db *gorm.DB, user *models.User, paginate *models.Pagination, onl
 			Order("DATE_TRUNC('year', date_shot) DESC").
 			Order("DATE_TRUNC('month', date_shot) DESC").
 			Order("DATE_TRUNC('day', date_shot) DESC").
-			Order("albums.title ASC").
+			Order(albumsTitleASC).
 			Order("media.date_shot DESC")
 	case drivers.SQLITE:
 		query = query.
 			Order("strftime('%Y-%m-%d', media.date_shot) DESC"). // convert to YYYY-MM-DD
-			Order("albums.title ASC").
+			Order(albumsTitleASC).
 			Order("TIME(media.date_shot) DESC")
 	default:
 		query = query.
 			Order("YEAR(media.date_shot) DESC").
 			Order("MONTH(media.date_shot) DESC").
 			Order("DAY(media.date_shot) DESC").
-			Order("albums.title ASC").
+			Order(albumsTitleASC).
 			Order("TIME(media.date_shot) DESC")
 	}
 
@@ -41,7 +44,11 @@ func MyTimeline(db *gorm.DB, user *models.User, paginate *models.Pagination, onl
 	}
 
 	if onlyFavorites != nil && *onlyFavorites {
-		query = query.Where("media.id IN (?)", db.Table("user_media_data").Select("user_media_data.media_id").Where("user_media_data.user_id = ?", user.ID).Where("user_media_data.favorite"))
+		query = query.
+			Where("media.id IN (?)", db.Table("user_media_data").
+				Select("user_media_data.media_id").
+				Where("user_media_data.user_id = ?", user.ID).
+				Where("user_media_data.favorite"))
 	}
 
 	query = models.FormatSQL(query, nil, paginate)
