@@ -23,32 +23,27 @@ func DeleteUser(db *gorm.DB, userID int) (*models.User, error) {
 	var user models.User
 	deletedAlbumIDs := make([]int, 0)
 
-	err := db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.First(&user, userID).Error; err != nil {
+	var err error
+	err = db.Transaction(func(tx *gorm.DB) error {
+		if err = tx.First(&user, userID).Error; err != nil {
 			return err
 		}
 
 		userAlbums := user.Albums
-		if err := tx.Model(&user).Association("Albums").Find(&userAlbums); err != nil {
+		if err = tx.Model(&user).Association("Albums").Find(&userAlbums); err != nil {
 			return err
 		}
 
-		if err := tx.Model(&user).Association("Albums").Clear(); err != nil {
+		if err = tx.Model(&user).Association("Albums").Clear(); err != nil {
 			return err
 		}
 
-		for _, album := range userAlbums {
-			var associatedUsers = tx.Model(album).Association("Owners").Count()
-
-			if associatedUsers == 0 {
-				deletedAlbumIDs = append(deletedAlbumIDs, album.ID)
-				if err := tx.Delete(album).Error; err != nil {
-					return err
-				}
-			}
+		deletedAlbumIDs, err = deleteNotOwnedAlbums(userAlbums, tx, deletedAlbumIDs)
+		if err != nil {
+			return err
 		}
 
-		if err := tx.Delete(&user).Error; err != nil {
+		if err = tx.Delete(&user).Error; err != nil {
 			return err
 		}
 
@@ -60,11 +55,30 @@ func DeleteUser(db *gorm.DB, userID int) (*models.User, error) {
 	}
 
 	// If there is only one associated user, clean up the cache folder and delete the album row
+	return &user, cleanup(deletedAlbumIDs)
+}
+
+func cleanup(deletedAlbumIDs []int) error {
+	var err error
 	for _, deletedAlbumID := range deletedAlbumIDs {
 		cachePath := path.Join(utils.MediaCachePath(), strconv.Itoa(int(deletedAlbumID)))
-		if err := os.RemoveAll(cachePath); err != nil {
-			return &user, err
+		if err = os.RemoveAll(cachePath); err != nil {
+			return err
 		}
 	}
-	return &user, nil
+	return err
+}
+
+func deleteNotOwnedAlbums(userAlbums []models.Album, tx *gorm.DB, deletedAlbumIDs []int) ([]int, error) {
+	for _, album := range userAlbums {
+		var associatedUsers = tx.Model(album).Association("Owners").Count()
+
+		if associatedUsers == 0 {
+			deletedAlbumIDs = append(deletedAlbumIDs, album.ID)
+			if err := tx.Delete(album).Error; err != nil {
+				return nil, err
+			}
+		}
+	}
+	return deletedAlbumIDs, nil
 }
