@@ -1,60 +1,22 @@
 package resolvers
 
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+// Code generated by github.com/99designs/gqlgen version v0.17.57
+
 import (
 	"context"
+	"errors"
 
 	api "github.com/photoview/photoview/api/graphql"
 	"github.com/photoview/photoview/api/graphql/auth"
 	"github.com/photoview/photoview/api/graphql/models"
 	"github.com/photoview/photoview/api/scanner/face_detection"
-	"github.com/pkg/errors"
 	"gorm.io/gorm"
 )
 
-type imageFaceResolver struct {
-	*Resolver
-}
-
-type faceGroupResolver struct {
-	*Resolver
-}
-
-func (r *Resolver) ImageFace() api.ImageFaceResolver {
-	return imageFaceResolver{r}
-}
-
-func (r *Resolver) FaceGroup() api.FaceGroupResolver {
-	return faceGroupResolver{r}
-}
-
-func (r imageFaceResolver) FaceGroup(ctx context.Context, obj *models.ImageFace) (*models.FaceGroup, error) {
-	if obj.FaceGroup != nil {
-		return obj.FaceGroup, nil
-	}
-
-	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
-	}
-
-	var faceGroup models.FaceGroup
-	if err := r.DB(ctx).Model(&obj).Association("FaceGroup").Find(&faceGroup); err != nil {
-		return nil, err
-	}
-
-	obj.FaceGroup = &faceGroup
-
-	return &faceGroup, nil
-}
-
-func (r imageFaceResolver) Media(ctx context.Context, obj *models.ImageFace) (*models.Media, error) {
-	if err := obj.FillMedia(r.DB(ctx)); err != nil {
-		return nil, err
-	}
-
-	return &obj.Media, nil
-}
-
-func (r faceGroupResolver) ImageFaces(ctx context.Context, obj *models.FaceGroup, paginate *models.Pagination) ([]*models.ImageFace, error) {
+// ImageFaces is the resolver for the imageFaces field.
+func (r *faceGroupResolver) ImageFaces(ctx context.Context, obj *models.FaceGroup, paginate *models.Pagination) ([]*models.ImageFace, error) {
 	db := r.DB(ctx)
 	user := auth.UserFromContext(ctx)
 	if user == nil {
@@ -62,7 +24,7 @@ func (r faceGroupResolver) ImageFaces(ctx context.Context, obj *models.FaceGroup
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
+		return nil, ErrFaceDetectorNotInitialized
 	}
 
 	if err := user.FillAlbums(db); err != nil {
@@ -76,7 +38,7 @@ func (r faceGroupResolver) ImageFaces(ctx context.Context, obj *models.FaceGroup
 
 	query := db.
 		Joins("Media").
-		Where("face_group_id = ?", obj.ID).
+		Where(faceGroupIDIsQuestion, obj.ID).
 		Where("album_id IN (?)", userAlbumIDs)
 
 	query = models.FormatSQL(query, nil, paginate)
@@ -89,7 +51,8 @@ func (r faceGroupResolver) ImageFaces(ctx context.Context, obj *models.FaceGroup
 	return imageFaces, nil
 }
 
-func (r faceGroupResolver) ImageFaceCount(ctx context.Context, obj *models.FaceGroup) (int, error) {
+// ImageFaceCount is the resolver for the imageFaceCount field.
+func (r *faceGroupResolver) ImageFaceCount(ctx context.Context, obj *models.FaceGroup) (int, error) {
 	db := r.DB(ctx)
 	user := auth.UserFromContext(ctx)
 	if user == nil {
@@ -97,7 +60,7 @@ func (r faceGroupResolver) ImageFaceCount(ctx context.Context, obj *models.FaceG
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return -1, errors.New("face detector not initialized")
+		return -1, ErrFaceDetectorNotInitialized
 	}
 
 	if err := user.FillAlbums(db); err != nil {
@@ -112,7 +75,7 @@ func (r faceGroupResolver) ImageFaceCount(ctx context.Context, obj *models.FaceG
 	query := db.
 		Model(&models.ImageFace{}).
 		Joins("Media").
-		Where("face_group_id = ?", obj.ID).
+		Where(faceGroupIDIsQuestion, obj.ID).
 		Where("album_id IN (?)", userAlbumIDs)
 
 	var count int64
@@ -123,78 +86,36 @@ func (r faceGroupResolver) ImageFaceCount(ctx context.Context, obj *models.FaceG
 	return int(count), nil
 }
 
-func (r *queryResolver) FaceGroup(ctx context.Context, id int) (*models.FaceGroup, error) {
-	db := r.DB(ctx)
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, errors.New("unauthorized")
+// Media is the resolver for the media field.
+func (r *imageFaceResolver) Media(ctx context.Context, obj *models.ImageFace) (*models.Media, error) {
+	if err := obj.FillMedia(r.DB(ctx)); err != nil {
+		return nil, err
+	}
+
+	return &obj.Media, nil
+}
+
+// FaceGroup is the resolver for the faceGroup field.
+func (r *imageFaceResolver) FaceGroup(ctx context.Context, obj *models.ImageFace) (*models.FaceGroup, error) {
+	if obj.FaceGroup != nil {
+		return obj.FaceGroup, nil
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
+		return nil, ErrFaceDetectorNotInitialized
 	}
-
-	if err := user.FillAlbums(db); err != nil {
-		return nil, err
-	}
-
-	userAlbumIDs := make([]int, len(user.Albums))
-	for i, album := range user.Albums {
-		userAlbumIDs[i] = album.ID
-	}
-
-	faceGroupQuery := db.
-		Joins("LEFT JOIN image_faces ON image_faces.face_group_id = face_groups.id").
-		Joins("LEFT JOIN media ON image_faces.media_id = media.id").
-		Where("face_groups.id = ?", id).
-		Where("media.album_id IN (?)", userAlbumIDs)
 
 	var faceGroup models.FaceGroup
-	if err := faceGroupQuery.Find(&faceGroup).Error; err != nil {
+	if err := r.DB(ctx).Model(&obj).Association("FaceGroup").Find(&faceGroup); err != nil {
 		return nil, err
 	}
+
+	obj.FaceGroup = &faceGroup
 
 	return &faceGroup, nil
 }
 
-func (r *queryResolver) MyFaceGroups(ctx context.Context, paginate *models.Pagination) ([]*models.FaceGroup, error) {
-	db := r.DB(ctx)
-	user := auth.UserFromContext(ctx)
-	if user == nil {
-		return nil, errors.New("unauthorized")
-	}
-
-	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
-	}
-
-	if err := user.FillAlbums(db); err != nil {
-		return nil, err
-	}
-
-	userAlbumIDs := make([]int, len(user.Albums))
-	for i, album := range user.Albums {
-		userAlbumIDs[i] = album.ID
-	}
-
-	faceGroupQuery := db.
-		Joins("JOIN image_faces ON image_faces.face_group_id = face_groups.id").
-		Where("image_faces.media_id IN (?)", db.Select("media.id").Table("media").Where("media.album_id IN (?)", userAlbumIDs)).
-		Group("image_faces.face_group_id").
-		Group("face_groups.id").
-		Order("CASE WHEN label IS NULL THEN 1 ELSE 0 END").
-		Order("COUNT(image_faces.id) DESC")
-
-	faceGroupQuery = models.FormatSQL(faceGroupQuery, nil, paginate)
-
-	var faceGroups []*models.FaceGroup
-	if err := faceGroupQuery.Find(&faceGroups).Error; err != nil {
-		return nil, err
-	}
-
-	return faceGroups, nil
-}
-
+// SetFaceGroupLabel is the resolver for the setFaceGroupLabel field.
 func (r *mutationResolver) SetFaceGroupLabel(ctx context.Context, faceGroupID int, label *string) (*models.FaceGroup, error) {
 	db := r.DB(ctx)
 	user := auth.UserFromContext(ctx)
@@ -203,7 +124,7 @@ func (r *mutationResolver) SetFaceGroupLabel(ctx context.Context, faceGroupID in
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
+		return nil, ErrFaceDetectorNotInitialized
 	}
 
 	faceGroup, err := userOwnedFaceGroup(db, user, faceGroupID)
@@ -218,6 +139,7 @@ func (r *mutationResolver) SetFaceGroupLabel(ctx context.Context, faceGroupID in
 	return faceGroup, nil
 }
 
+// CombineFaceGroups is the resolver for the combineFaceGroups field.
 func (r *mutationResolver) CombineFaceGroups(ctx context.Context, destinationFaceGroupID int, sourceFaceGroupID int) (*models.FaceGroup, error) {
 	db := r.DB(ctx)
 	user := auth.UserFromContext(ctx)
@@ -226,7 +148,7 @@ func (r *mutationResolver) CombineFaceGroups(ctx context.Context, destinationFac
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
+		return nil, ErrFaceDetectorNotInitialized
 	}
 
 	destinationFaceGroup, err := userOwnedFaceGroup(db, user, destinationFaceGroupID)
@@ -240,7 +162,9 @@ func (r *mutationResolver) CombineFaceGroups(ctx context.Context, destinationFac
 	}
 
 	updateError := db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Model(&models.ImageFace{}).Where("face_group_id = ?", sourceFaceGroup.ID).Update("face_group_id", destinationFaceGroup.ID).Error; err != nil {
+		if err := tx.Model(&models.ImageFace{}).
+			Where(faceGroupIDIsQuestion, sourceFaceGroup.ID).
+			Update("face_group_id", destinationFaceGroup.ID).Error; err != nil {
 			return err
 		}
 
@@ -260,6 +184,7 @@ func (r *mutationResolver) CombineFaceGroups(ctx context.Context, destinationFac
 	return destinationFaceGroup, nil
 }
 
+// MoveImageFaces is the resolver for the moveImageFaces field.
 func (r *mutationResolver) MoveImageFaces(ctx context.Context, imageFaceIDs []int, destinationFaceGroupID int) (*models.FaceGroup, error) {
 	db := r.DB(ctx)
 	user := auth.UserFromContext(ctx)
@@ -268,7 +193,7 @@ func (r *mutationResolver) MoveImageFaces(ctx context.Context, imageFaceIDs []in
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
+		return nil, ErrFaceDetectorNotInitialized
 	}
 
 	userOwnedImageFaceIDs := make([]int, 0)
@@ -294,7 +219,7 @@ func (r *mutationResolver) MoveImageFaces(ctx context.Context, imageFaceIDs []in
 		var sourceFaceGroups []*models.FaceGroup
 		if err := tx.
 			Joins("LEFT JOIN image_faces ON image_faces.face_group_id = face_groups.id").
-			Where("image_faces.id IN (?)", userOwnedImageFaceIDs).
+			Where(imageFacesIDInQuestion, userOwnedImageFaceIDs).
 			Find(&sourceFaceGroups).Error; err != nil {
 			return err
 		}
@@ -307,17 +232,8 @@ func (r *mutationResolver) MoveImageFaces(ctx context.Context, imageFaceIDs []in
 		}
 
 		// delete face groups if they have become empty
-		for _, faceGroup := range sourceFaceGroups {
-			var count int64
-			if err := tx.Model(&models.ImageFace{}).Where("face_group_id = ?", faceGroup.ID).Count(&count).Error; err != nil {
-				return err
-			}
-
-			if count == 0 {
-				if err := tx.Delete(&faceGroup).Error; err != nil {
-					return err
-				}
-			}
+		if err := deleteEmptyFaceGroups(sourceFaceGroups, tx); err != nil {
+			return err
 		}
 
 		return nil
@@ -332,6 +248,7 @@ func (r *mutationResolver) MoveImageFaces(ctx context.Context, imageFaceIDs []in
 	return destFaceGroup, nil
 }
 
+// RecognizeUnlabeledFaces is the resolver for the recognizeUnlabeledFaces field.
 func (r *mutationResolver) RecognizeUnlabeledFaces(ctx context.Context) ([]*models.ImageFace, error) {
 	db := r.DB(ctx)
 	user := auth.UserFromContext(ctx)
@@ -340,7 +257,7 @@ func (r *mutationResolver) RecognizeUnlabeledFaces(ctx context.Context) ([]*mode
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
+		return nil, ErrFaceDetectorNotInitialized
 	}
 
 	var updatedImageFaces []*models.ImageFace
@@ -359,6 +276,7 @@ func (r *mutationResolver) RecognizeUnlabeledFaces(ctx context.Context) ([]*mode
 	return updatedImageFaces, nil
 }
 
+// DetachImageFaces is the resolver for the detachImageFaces field.
 func (r *mutationResolver) DetachImageFaces(ctx context.Context, imageFaceIDs []int) (*models.FaceGroup, error) {
 	db := r.DB(ctx)
 	user := auth.UserFromContext(ctx)
@@ -367,7 +285,7 @@ func (r *mutationResolver) DetachImageFaces(ctx context.Context, imageFaceIDs []
 	}
 
 	if face_detection.GlobalFaceDetector == nil {
-		return nil, errors.New("face detector not initialized")
+		return nil, ErrFaceDetectorNotInitialized
 	}
 
 	userOwnedImageFaceIDs := make([]int, 0)
@@ -407,14 +325,16 @@ func (r *mutationResolver) DetachImageFaces(ctx context.Context, imageFaceIDs []
 	return &newFaceGroup, nil
 }
 
-func userOwnedFaceGroup(db *gorm.DB, user *models.User, faceGroupID int) (*models.FaceGroup, error) {
-	if user.Admin {
-		var faceGroup models.FaceGroup
-		if err := db.Where("id = ?", faceGroupID).Find(&faceGroup).Error; err != nil {
-			return nil, err
-		}
+// MyFaceGroups is the resolver for the myFaceGroups field.
+func (r *queryResolver) MyFaceGroups(ctx context.Context, paginate *models.Pagination) ([]*models.FaceGroup, error) {
+	db := r.DB(ctx)
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, errors.New("unauthorized")
+	}
 
-		return &faceGroup, nil
+	if face_detection.GlobalFaceDetector == nil {
+		return nil, ErrFaceDetectorNotInitialized
 	}
 
 	if err := user.FillAlbums(db); err != nil {
@@ -426,32 +346,38 @@ func userOwnedFaceGroup(db *gorm.DB, user *models.User, faceGroupID int) (*model
 		userAlbumIDs[i] = album.ID
 	}
 
-	// Verify that user owns at leat one of the images in the face group
-	imageFaceQuery := db.
-		Select("image_faces.id").
-		Table("image_faces").
-		Joins("JOIN media ON media.id = image_faces.media_id").
-		Where("media.album_id IN (?)", userAlbumIDs)
-
 	faceGroupQuery := db.
-		Model(&models.FaceGroup{}).
-		Joins("JOIN image_faces ON face_groups.id = image_faces.face_group_id").
-		Where("face_groups.id = ?", faceGroupID).
-		Where("image_faces.id IN (?)", imageFaceQuery)
+		Joins("JOIN image_faces ON image_faces.face_group_id = face_groups.id").
+		Where("image_faces.media_id IN (?)",
+			db.Select("media.id").Table("media").Where(mediaAlbumIDInQuestion, userAlbumIDs)).
+		Group("image_faces.face_group_id").
+		Group("face_groups.id").
+		Order("CASE WHEN label IS NULL THEN 1 ELSE 0 END").
+		Order("COUNT(image_faces.id) DESC")
 
-	var faceGroup models.FaceGroup
-	if err := faceGroupQuery.Find(&faceGroup).Error; err != nil {
-		if err == gorm.ErrRecordNotFound {
-			return nil, errors.Wrap(err, "face group does not exist or is not owned by the user")
-		}
+	faceGroupQuery = models.FormatSQL(faceGroupQuery, nil, paginate)
+
+	var faceGroups []*models.FaceGroup
+	if err := faceGroupQuery.Find(&faceGroups).Error; err != nil {
 		return nil, err
 	}
 
-	return &faceGroup, nil
+	return faceGroups, nil
 }
 
-func getUserOwnedImageFaces(tx *gorm.DB, user *models.User, imageFaceIDs []int) ([]*models.ImageFace, error) {
-	if err := user.FillAlbums(tx); err != nil {
+// FaceGroup is the resolver for the faceGroup field.
+func (r *queryResolver) FaceGroup(ctx context.Context, id int) (*models.FaceGroup, error) {
+	db := r.DB(ctx)
+	user := auth.UserFromContext(ctx)
+	if user == nil {
+		return nil, errors.New("unauthorized")
+	}
+
+	if face_detection.GlobalFaceDetector == nil {
+		return nil, ErrFaceDetectorNotInitialized
+	}
+
+	if err := user.FillAlbums(db); err != nil {
 		return nil, err
 	}
 
@@ -460,14 +386,25 @@ func getUserOwnedImageFaces(tx *gorm.DB, user *models.User, imageFaceIDs []int) 
 		userAlbumIDs[i] = album.ID
 	}
 
-	var userOwnedImageFaces []*models.ImageFace
-	if err := tx.
-		Joins("JOIN media ON media.id = image_faces.media_id").
-		Where("media.album_id IN (?)", userAlbumIDs).
-		Where("image_faces.id IN (?)", imageFaceIDs).
-		Find(&userOwnedImageFaces).Error; err != nil {
+	faceGroupQuery := db.
+		Joins("LEFT JOIN image_faces ON image_faces.face_group_id = face_groups.id").
+		Joins("LEFT JOIN media ON image_faces.media_id = media.id").
+		Where("face_groups.id = ?", id).
+		Where(mediaAlbumIDInQuestion, userAlbumIDs)
+
+	var faceGroup models.FaceGroup
+	if err := faceGroupQuery.Find(&faceGroup).Error; err != nil {
 		return nil, err
 	}
 
-	return userOwnedImageFaces, nil
+	return &faceGroup, nil
 }
+
+// FaceGroup returns api.FaceGroupResolver implementation.
+func (r *Resolver) FaceGroup() api.FaceGroupResolver { return &faceGroupResolver{r} }
+
+// ImageFace returns api.ImageFaceResolver implementation.
+func (r *Resolver) ImageFace() api.ImageFaceResolver { return &imageFaceResolver{r} }
+
+type faceGroupResolver struct{ *Resolver }
+type imageFaceResolver struct{ *Resolver }
