@@ -1,47 +1,18 @@
 package resolvers
 
-import (
-	"context"
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+// Code generated by github.com/99designs/gqlgen version v0.17.57
 
+import (
 	api "github.com/photoview/photoview/api/graphql"
-	"gorm.io/gorm"
 )
 
-//go:generate go run github.com/99designs/gqlgen
+// Mutation returns api.MutationResolver implementation.
+func (r *Resolver) Mutation() api.MutationResolver { return &mutationResolver{r} }
 
-type Resolver struct {
-	database *gorm.DB
-}
-
-func NewRootResolver(db *gorm.DB) Resolver {
-	return Resolver{
-		database: db,
-	}
-}
-
-// DB returns a database instance that is tied to the given context
-func (r *Resolver) DB(ctx context.Context) *gorm.DB {
-	return r.database.WithContext(ctx)
-}
-
-func (r *Resolver) Mutation() api.MutationResolver {
-	return &mutationResolver{r}
-}
-
-func (r *Resolver) Query() api.QueryResolver {
-	return &queryResolver{r}
-}
-
-func (r *Resolver) Subscription() api.SubscriptionResolver {
-	return &subscriptionResolver{
-		Resolver: r,
-	}
-}
+// Query returns api.QueryResolver implementation.
+func (r *Resolver) Query() api.QueryResolver { return &queryResolver{r} }
 
 type mutationResolver struct{ *Resolver }
-
 type queryResolver struct{ *Resolver }
-
-type subscriptionResolver struct {
-	Resolver *Resolver
-}
