@@ -1,13 +1,17 @@
 package resolvers
 
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+// Code generated by github.com/99designs/gqlgen version v0.17.57
+
 import (
 	"context"
 
 	"github.com/photoview/photoview/api/graphql/models"
-	// "github.com/pkg/errors"
 	"gorm.io/gorm"
 )
 
+// SetThumbnailDownsampleMethod is the resolver for the setThumbnailDownsampleMethod field.
 func (r *mutationResolver) SetThumbnailDownsampleMethod(ctx context.Context, method models.ThumbnailFilter) (models.ThumbnailFilter, error) {
 	db := r.DB(ctx)
 
@@ -15,7 +19,12 @@ func (r *mutationResolver) SetThumbnailDownsampleMethod(ctx context.Context, met
 	// 	return 0, errors.New("The requested filter is unsupported, defaulting to nearest neighbor")
 	// }
 
-	if err := db.Session(&gorm.Session{AllowGlobalUpdate: true}).Model(&models.SiteInfo{}).Update("thumbnail_method", method).Error; err != nil {
+	if err := db.
+		Session(&gorm.Session{AllowGlobalUpdate: true}).
+		Model(&models.SiteInfo{}).
+		Update("thumbnail_method", method).
+		Error; err != nil {
+
 		return models.ThumbnailFilterNearestNeighbor, err
 	}
 
@@ -31,6 +40,4 @@ func (r *mutationResolver) SetThumbnailDownsampleMethod(ctx context.Context, met
 	// 	lng := models.LanguageTranslation(*language)
 	// 	langTrans = &lng
 	// }
-
-
 }
