@@ -43,11 +43,12 @@ func TestExifParsers(t *testing.T) {
 
 	images := []struct {
 		path   string
-		assert func(t *testing.T, exif *models.MediaEXIF)
+		assert func(t *testing.T, exif *models.MediaEXIF, err error)
 	}{
 		{
 			path: "./test_data/bird.jpg",
-			assert: func(t *testing.T, exif *models.MediaEXIF) {
+			assert: func(t *testing.T, exif *models.MediaEXIF, err error) {
+				assert.NoError(t, err)
 				assert.EqualValues(t, *exif.Description, "Photo of a Bird")
 				assert.WithinDuration(t, *exif.DateShot, time.Unix(1336318784, 0).UTC(), time.Minute)
 				assert.EqualValues(t, *exif.Camera, "Canon EOS 600D")
@@ -65,14 +66,59 @@ func TestExifParsers(t *testing.T) {
 		},
 		{
 			path: "./test_data/stripped.jpg",
-			assert: func(t *testing.T, exif *models.MediaEXIF) {
-				assert.Nil(t, exif)
+			assert: func(t *testing.T, exif *models.MediaEXIF, err error) {
+				assert.NoError(t, err)
+				if exif == nil {
+					assert.Nil(t, exif)
+				} else {
+					assert.Equal(t, 0, exif.ID)
+					assert.True(t, exif.CreatedAt.IsZero())
+					assert.True(t, exif.UpdatedAt.IsZero())
+					assert.Nil(t, exif.Description)
+					assert.Nil(t, exif.Camera)
+					assert.Nil(t, exif.Maker)
+					assert.Nil(t, exif.Lens)
+					assert.Nil(t, exif.Exposure)
+					assert.Nil(t, exif.Aperture)
+					assert.Nil(t, exif.Iso)
+					assert.Nil(t, exif.FocalLength)
+					assert.Nil(t, exif.Flash)
+					assert.Nil(t, exif.Orientation)
+					assert.Nil(t, exif.ExposureProgram)
+					assert.Nil(t, exif.GPSLatitude)
+					assert.Nil(t, exif.GPSLongitude)
+				}
 			},
 		},
 		{
 			path: "./test_data/bad-exif.jpg",
-			assert: func(t *testing.T, exif *models.MediaEXIF) {
+			assert: func(t *testing.T, exif *models.MediaEXIF, err error) {
+				assert.NoError(t, err)
 				assert.Nil(t, exif.Exposure)
+			},
+		},
+		{
+			path: "./test_data/IncorrectGPS.jpg",
+			assert: func(t *testing.T, exif *models.MediaEXIF, err error) {
+				assert.Nil(t, exif.GPSLatitude,
+					"GPSLatitude expected to be NULL for an incorrect input data: %+v", exif.GPSLatitude)
+				assert.Nil(t, exif.GPSLongitude,
+					"GPSLongitude expected to be NULL for an incorrect input data: %+v", exif.GPSLongitude)
+			},
+		},
+		{
+			path: "./test_data/CorrectGPS.jpg",
+			assert: func(t *testing.T, exif *models.MediaEXIF, err error) {
+				const precision = 1e-7
+				assert.NoError(t, err)
+				assert.NotNil(t, exif.GPSLatitude,
+					"GPSLatitude expected to be Not-NULL for a correct input data: %+v", exif.GPSLatitude)
+				assert.NotNil(t, exif.GPSLongitude,
+					"GPSLongitude expected to be Not-NULL for a correct input data: %+v", exif.GPSLongitude)
+				assert.InDelta(t, *exif.GPSLatitude, 44.478997222222226, precision,
+					"The exact value from input data is expected: %+v", exif.GPSLatitude)
+				assert.InDelta(t, *exif.GPSLongitude, 11.297922222222223, precision,
+					"The exact value from input data is expected: %+v", exif.GPSLongitude)
 			},
 		},
 	}
@@ -90,9 +136,7 @@ func TestExifParsers(t *testing.T) {
 
 				exif, err := p.parser.ParseExif(img.path)
 
-				if assert.NoError(t, err) {
-					img.assert(t, exif)
-				}
+				img.assert(t, exif, err)
 			})
 		}
 	}
