@@ -22,6 +22,7 @@ const (
 	TypeWebp MediaType = "image/webp"
 	TypeBmp  MediaType = "image/bmp"
 	TypeHeic MediaType = "image/heic"
+	TypeGif  MediaType = "image/gif"
 
 	// Raw formats
 	TypeDNG MediaType = "image/x-adobe-dng"
@@ -78,6 +79,7 @@ var SupportedMimetypes = [...]MediaType{
 	TypeWebp,
 	TypeBmp,
 	TypeHeic,
+	TypeGif,
 }
 
 var WebMimetypes = [...]MediaType{
@@ -85,6 +87,7 @@ var WebMimetypes = [...]MediaType{
 	TypePng,
 	TypeWebp,
 	TypeBmp,
+	TypeGif,
 }
 
 var RawMimeTypes = [...]MediaType{
@@ -152,6 +155,7 @@ var fileExtensions = map[string]MediaType{
 	".tiff": TypeTiff,
 	".bmp":  TypeBmp,
 	".heic": TypeHeic,
+	".gif":  TypeGif,
 
 	// RAW formats
 	".dng": TypeDNG,
@@ -256,11 +260,11 @@ func (imgType *MediaType) IsSupported() bool {
 		return true
 	}
 
-	if executable_worker.DarktableCli.IsInstalled() && imgType.IsRaw() {
+	if executable_worker.Magick.IsInstalled() && imgType.IsRaw() {
 		return true
 	}
 
-	if executable_worker.FfmpegCli.IsInstalled() && imgType.IsVideo() {
+	if executable_worker.Ffmpeg.IsInstalled() && imgType.IsVideo() {
 		return true
 	}
 
