@@ -71,17 +71,17 @@ func ScanMedia(tx *gorm.DB, mediaPath string, albumId int, cache *scanner_cache.
 // ProcessSingleMedia processes a single media, might be used to reprocess media with corrupted cache
 // Function waits for processing to finish before returning.
 func ProcessSingleMedia(db *gorm.DB, media *models.Media) error {
-	album_cache := scanner_cache.MakeAlbumCache()
+	albumCache := scanner_cache.MakeAlbumCache()
 
 	var album models.Album
 	if err := db.Model(media).Association("Album").Find(&album); err != nil {
 		return err
 	}
 
-	media_data := media_encoding.NewEncodeMediaData(media)
+	mediaData := media_encoding.NewEncodeMediaData(media)
 
-	task_context := scanner_task.NewTaskContext(context.Background(), db, &album, album_cache)
-	if err := scanMedia(task_context, media, &media_data, 0, 1); err != nil {
+	taskContext := scanner_task.NewTaskContext(context.Background(), db, &album, albumCache)
+	if err := scanMedia(taskContext, media, &mediaData, 0, 1); err != nil {
 		return errors.Wrap(err, "single media scan")
 	}
 
