@@ -10,10 +10,11 @@ type MediaCleanupTask struct {
 	scanner_task.ScannerTaskBase
 }
 
-func (t MediaCleanupTask) AfterScanAlbum(ctx scanner_task.TaskContext, changedMedia []*models.Media, albumMedia []*models.Media) error {
+func (t MediaCleanupTask) AfterScanAlbum(ctx scanner_task.TaskContext, changedMedia []*models.Media,
+	albumMedia []*models.Media) error {
 
-	cleanup_errors := CleanupMedia(ctx.GetDB(), ctx.GetAlbum().ID, albumMedia)
-	for _, err := range cleanup_errors {
+	cleanupErrors := CleanupMedia(ctx.GetDB(), ctx.GetAlbum().ID, albumMedia)
+	for _, err := range cleanupErrors {
 		scanner_utils.ScannerError("delete old media: %s", err)
 	}
 
